requirements.txt
# Core web framework
flask>=3.0.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
